refactor: 역할에 따라 config 파일 분리
